Improved handling display of room name and messages with yellow background
Room names that are too long to fit on the screen are now displayed as
they are in VVVVVV (just centered with both sides cut off. When a message
with a yellow background is shown, the text can now span multiple lines.
When editing a room name, you can see multiple lines as well, but any
additional lines have a red background color.

Closes #1