Fixed jumping to script from (non-script editor) search feature
This used to simply leave the first line of the script in the input
buffers while just changing the number of the line that the cursor is on.

Bonus is that the line which is jumped to will now always be on the
screen immediately (that means, scrolled to) because I've changed it to
use scriptgotoline().

Closes #2