Made entity IDs and entity count more reliable
This fixes the bug where placing a script box, terminal or room text would
not let you type a script name or room text, and would create a 0x0 script
box or terminal with the script name "stop", or a roomtext "[Bug!]".

Now, instead of relying on what table.insert(table, newelement) and #table
do, an incrementing variable count.entity_ai is kept, comparable to
MySQL's AUTO_INCREMENT, to ensure consistency between what key an entity
was given, and what key Ved is made to think that entity was given.

Also fixed a bug where both placing the entrance and the exit of a warp
token were counted as new entities.