Added a pcall for the crash screen
Now if the crash screen crashes, Ved won't just close, but will display a
more basic crash screen.

This one not crashing is not guaranteed, though, so I /might/ need to add
another crash screen for when the crash screen crash screen crashes. Then
if the crash screen crash screen crash screen crashes, we need another one
that just displays a background color or something. If that doesn't even
work, we're gonna need to resort to using an available printer to print
the five error messages on paper, and hope the ink hasn't run out.