The XML handler now handles &#32; well (even when saving)
That is the XML entity that stands for a space.

In short: you can now use multiple spaces in a row, and you can use spaces
at the starts and ends of text.

When loading a level, &#32; is converted into a space, like it should.

When saving a level, any spaces at the start or end of strings, or
multiple spaces in a row, are converted so that VVVVVV's internal XML
parser doesn't remove them upon loading.