Made the code style more consistent
- Lines should now no longer consist of just indentation (whether on the
same level as surrounding lines or otherwise) because this was an
invisible chaos. Not that it mattered much, but it was really easy to
mass-change.
- Almost all (non-multiline) strings have been using double quotes instead
of apostrophes, the few strings using apostrophes now use quotes as well,
except for a few strings with Windows command line related strings which
already have quotes within them, because changing to quotes there would
make it less readable.