Ved now backups level files that are overwritten when saving
It's on by default, you can turn the feature off in the options. By
default, up to 5 backups are kept per level before the oldest backups
start getting deleted, which is also configurable, but that's not in the
options screen yet.

Backups are kept in Ved's save directory (the one that's inside %appdata%
on Windows), inside the folder called "overwrite_backups". Each level name
gets its own folder, and the names of the levels inside start with a
(UNIX) timestamp of when the level file was overwritten, followed by a
timestamp of when the overwritten file was last modified. That means a
backup called 2222222222_1111111111_levelname.vvvvvv was originally saved
at 1111111111 seconds since 1970, and overwritten by something else at
2222222222 seconds since 1970.

There will be a screen for displaying and managing all backups in Ved
later, which will also display these UNIX-timestamps in human-readable
dates.

For Windows, there doesn't seem to be any (remotely worthy) way of getting
a last modified time via the command line, so I made a small .exe to do
that.