Made scale and amount of backups configurable within Ved
Scaling finally uses -/+ buttons, and can be set to a value of 1-9. As a
bonus, the value will be red if it's higher than what fits on any of your
displays, and that is calculated dynamically.

The same control is used for changing the number of backups that are kept
for each level. NOTE: You can hold down shift to rapidly increase/decrease
the value, for example, if you want to change it from 5 to 150 backups,
you don't have to click 145 times.