Added proper escaping of filenames and other arguments (for all OSes)
It now assumes even the setting for a custom VVVVVV dir is entered by the
user maliciously, so you can no longer inject commands via the custom
VVVVVV folder setting. That makes this kind of a security fix as well,
although you would have had to deliberately cause any issues for yourself,
and setting the custom VVVVVV folder was the only way to do it.

(the same goes for the filename passed to either `stat` or `fileunix.exe`
when overwriting a level, but that is added in this version, so it has not
officially been released yet)