Fixed filefunc_win.lua:109: table index is nil if levels folder invalid
If, on Windows, the levels folder is either set to one with slashes
instead of backslashes or has a trailing (back)slash, then that will be
handled, and not cause a crash. If it even still fails in a similar way,
it won't crash, it will just skip that level, which may result in an empty
levels list. I should probably show an error in that case.