Updated the internal commands reference more
Some descriptions were changed and commands were added based on
suggestions, and I also added a few other commands to the list. Also
slightly changed the definition of the color coding - red commands now
don't have to fully destroy save data to be red, because I added a command
which unlocks the secret lab, which also negatively impacts the integrity
of the main game save.