Added partially transparent background to menus in main editor
That is, the menu on the right, and the small 32px wide bar on the left IF
in smaller screen mode. The normal menu on the left side (when not in
smaller screen mode, or when expanded in smaller screen mode) already had
this background, which normally (outside of smaller screen mode) wasn't
visible, because it's transparent black on black, except when using a
larger cursor. The aforementioned areas now have that background as well,
so that when using larger cursors (especially in manual mode, or when
stamping) it's better visible which tiles are outside of the room.

Also, apparently I accidentally added carriage returns to specifically the
license, which I haven't touched. ¯\_(ツ)_/¯