Added optimization for long articles in the help system (thus level notes)
Offscreen lines are not rendered anymore (except up to 1024 pixels above
the top of the screen, to allow for high images to not suddenly disappear
as you scroll down).

This makes the changelog note in a version of open2v2, which has more than
1600 lines and uses ¤ on most of them, render at 60 FPS for me, up from
about 32.

Next, I'll try to optimize the parsing of lines which have many different
colors on that one line, which also really needs to be done.