Added optimization for many colors on one line in help/level notes
There's now a cache for the different colored parts of each line, so the
possibly intensive task of splitting now no longer happens every time the
line is drawn.

This also very much optimizes the Tiles note in a version of open2v2 for
me, which uses more than a dozen different colored parts on several lines.
It used to be rendered at about 20 FPS for me, now it also does it at a
solid 60 FPS.

(Oh, and that's with vsync, if I turn vsync off then Changelog renders at
250-485 FPS, and Tiles renders at more than 300)