Command to get "originally last overwritten" timestamp now uses %appdata%
It used to use the full path. As described in commit
99d5cc6cef87d60b59b051f209ccd750350bf6e1, quoting the entire path to
prevent issues with spaces in the path is, bizarrely, invalid syntax via
io.popen, but if you leave them out, you're gonna have problems if your
username contains spaces, or if you happen to still be using Windows XP
(hello again, "Documents and Settings").

Using %appdata% solves this problem.