Added support for update notes and automatic saving of crash logs
Update notes are basically help pages that come along with the update
check, so it'll be possible to show things like the current changelog in
Ved. (And it'd be possible to show what has changed in versions newer than
yours, which would be really nice)

Automatic saving of crash logs saves information about crashes to file,
including and limited to the information that is displayed on the crash
screen, the time, whether debug mode is enabled, whether it was saved
because it was enabled or because the settings haven't been loaded yet,
and a traceback. This can be useful in case the crash screen crashes, the
traceback is needed, the info is needed after closing the Ved with the
crash screen, or for sending more info about a crash to me. Maybe I'll
implement a button to allow people to send a crash report to me if they
press it. Automatically saving crash logs to files can be disabled in the
settings.

Also, the crash screen now also displays the time since Ved was started,
which could be useful to get a quick impression of whether the error
occurred while starting up or later.