Implemented support for links in help/level notes
Links can link to:
- Other articles/notes
- Anchors (any line with the # formatting code) in either the same
  article/note or other articles/notes
- Websites

The displayed text for the link can be customized and colored, but when
hovering over it, the destination is shown.

This allows for complete index tables (I added an index table to the
"Lists reference" help article for all the different lists) and of course,
links to websites.

https://i.imgur.com/V7Glt2s.png

I'll type all the details and the ways of using it later