Added tracking of unsaved changes
The "Are you sure you want to quit?" dialog is now replaced by a dialog
that only shows up if there are unsaved changes. You can choose to save
the changes, discard them, or cancel. (If you save, you can still change
the filename to save as)

This uses the undo and redo stacks to know whether or not unsaved changes
were undone or redone.

I also added an asterisk indicator in the window title to indicate unsaved
changes.