Fixed bug where next click was ignored if using ctrl+right in script
mousepressed would be set to true as though you clicked on the button
to jump to the script, while there's no mouse button release, so the
next time you would try to place an entity or click many buttons, the
first click would be ignored.

Thanks to Info Teddy for reporting this.