Fixed input to right of cursor vanishing if opening dialogs in script
Opening a dialog with an input field in the script editor (like the one
to create a new script) while having moved the text cursor to a
position before the end of a line would cause the text to the right of
the cursor to disappear until moving the cursor again.

This is not the most elegant fix (now the text cursor is moved to the
end of the line before opening the dialog) because making text not
disappear while leaving the cursor in place would require a lot of the
text input code to be rewritten. That could be done later.

Closes #13.