Added s.vis_firstseen and s.visload_seen settings
These are already used in the april fools' playtesting mode plugins.
These settings will probably be removed later, and when they do get
removed, visload would show the "1.4 update" screen on every session,
and vis would show the "to be clear" on every session. I think adding
a way for plugins to add settings for themselves would be a good idea
though, so I'd probably implement that and update vis first!