Fixed config default arrays being set by reference
This meant that if an array type config option doesn't exist yet in the
settings file, and the value is changed, it would change the default
value in memory too.
This mostly doesn't cause problems, unless you want to use that default
again, like when resetting the syntax colors to default.

Classic mistake since config options were only numbers and strings back
then, and I have a feeling we've fixed this before in [\\], which has a
very similar configuration system.