First phase of improving dialog-related code
I've split dialog.lua up into multiple different files, to concentrate
dialog-related code in dialog.lua. The option not to show a title bar
on dialog boxes has already been removed for cleanup (the parameter is
still in the dialog.new function, but has no effect anymore). Added
dialog.is_open(), which is especially meant for use in the rest of the
code instead of DIAwindowani.

I'm also bumping the version number from 1.3.4 to 1.4.0.