Added majority of overhauled dialog code
Dialogs are getting a major overhaul internally, to allow multiple
on-screen dialogs at the same time, and to have a MUCH better API
and internals.

The quit confirmation and save dialogs are the first to use the new
system, so dialogs being on-screen won't block quitting anymore. The
plan is to migrate all dialogs to the new system and then to remove all
code belonging to the old one, but for a little while I'll probably
leave both systems working in parallel, which works pretty well.

Testing is much appreciated! Especially since saving is concerned here,
so should you notice anything weird going on - like changes being
discarded and Ved closing without having used the Discard button, the
save dialog silently not saving in certain cases, or anything else,
better if it's found out now than after the next release.

There's also a very temporary dev shortcut (F12) to spawn a new-style
dialog any time.

Closes #31.
Closes #33.