Added detailed flag usage checking feature, small dialogs change
TLDR: You can now see what scripts a flag was used in, and some
      technical changes.

When clicking on a flag on the flags screen, you can now also see in
what scripts that flag has been used, along with on which lines.
(I might remove the latter, because a flag may appear 200 times in
one script and you won't be able to see anything else)

Closes #20.

The flag name dialog has also been migrated to the new dialogs system,
and a few bugs were fixed (turns out an error being displayed if you
don't change a flag name and hit OK was a bug after all, due to a
typo in a variable name. Also, humorously, the flag lookup function was
treating every line in every script as a script on its own, that needed
to be split by newlines. Each of those scripts had 1 line in total, so
it had no effect. No idea how late it was when I wrote that, it was in
the initial commit.)

Also, the dialog system is changing a little bit: no-close checkers
(which I'm thinking of calling "validators" instead for less confusion)
should now spawn any error message dialogs, instead of the handler.
Otherwise, error validation would happen in two different functions.
Dialog handlers now get a fourth argument telling if the dialog has
been kept on screen, so they can simply return.