Flag usage checker now only lists unique script names, not every usage
A flag may be used 200 times in the same script, and in that case you
won't see a massive (cut-off) list of that same script. You can now
only see unique script names where a flag was used, but a number of
usages is still displayed which lists how many times a flag was used
in total.