Current level (in title) now only gets updated on successful save/load
Causing an error by trying to save a title that is too long for the
filesystem, or trying to overwrite a read-only level, would still
change the current level name in the title bar. Likewise, trying to
load a level when impossible to do so (no such file, file corrupted)
would change the title if you didn't already have a level open.