Added variable storage to metadata entity
It's now possible (for plugins, or Ved in the future) to store
variables in levels, in the metadata entity.

The following functions can be used to interface with this:

get_level_var(key)
get_all_level_vars()
set_level_var(key, value)
remove_level_var(key)

Right now, the types of values that are supported are strings, numbers
and booleans.

This bumps the metadata entity version number to version 3.