New dialogs now support dropdowns
This works a bit different too, I'll still write documentation about
this. Basically, instead of providing a "right click menu id", you can
now provide a function that may determine what the value of the field
is set to.

For reference, element 7 is a table listing the items in the dropdown
menu, element 8 is an optional table that converts a value to a
displayable "current selection" if you want to pass the value in
another way than the displayed string (in which case you should provide
the aforementioned function so that you can change the selected item
text into a value that is a key in element 8. Did I mention that if you
don't want to provide element 8, you have to make it `false`, otherwise
Lua will consider the table terminated?

(anyway, the documentation will come soon™)