Tile solidity/spikyness can now be checked by holding J
If you hold J, tiles which are solid will have a white cross on them,
and tiles which are spikes are displayed with a red cross on them, so
you can clearly see which tiles are solid, which aren't, and which are
spikes, even if the tileset doesn't make that clear. This works both
on normal room contents and on the tiles picker.

In addition, holding N to see tile numbers now also works on the tiles
picker.