Map is now loaded at different priority depending on state
If you're looking at the map itself, more time per frame will be
allowed for it. If you're looking at, for example, the script editor,
the map will still be loaded, but allowing less time per frame for
doing so, so the map will be loaded at a slower rate.