Moved scissor disabling trick for first rooms on map
(Probably boring technical stuff)

Turns out that bug where the first two or three rooms on the map would
be blank after loading a big level was my fault, the scissor came from
the levels list, not the button for tool 9 in the main editor or
anything.
Basically, the bug was caused by the first rooms for the map being
drawn before a frame of the main editor had been drawn. The levels list
scissors the "recently used levels" box, and when you load a level
from there, the rest of the drawing code is skipped, including
disabling the scissor. The main editor uses scissors as well so this
will go visually unnoticed since it's going to be overwritten.

My original fix was to disable the scissor before drawing each map
room, but it can just be done once when loading the level.