Made Int.sc mode use squeak(off) and text(1,0,0,3)
Since VVVVVV converts all custom scripts to internal scripts, and these
converted scripts have a script lines limit of 500 lines, too, it would
be helpful to make sure all converted scripts have as little lines as
possible.

In this case, we are removing the `squeak(terminal)` the game generates
for each `say` block by adding a `squeak(off)` at the top of each
script. This saves 1 line per `say` block. This means we are overwriting
one less line, so we use `text(1,0,0,3)` instead.

The old format of using no `squeak(off)` and using `text(1,0,0,4)` can
still be read, but once a script with that format is closed, it is
converted to the new format.

Closes #41.