Migrated even more dialogs to the new dialog system
Also discovered that attempting to make a load script with a new flag
when all flags have been used resulted in a message as if you're in the
script editor, and used "FLAG" as a substitute. Now it creates a
loadscript with iftrinkets(0,...) instead, and gives a more appropriate
message.