Updated translation tools for separate main/help projects
Behind the scenes: help pages were in a separate file in the same
project in Pootle, but this was a bit annoying since the help pages
were sorted before the main strings, were annoying to scroll through
in search results, obscured how much progress was made in the main
file, and some other things. Now I made the help pages a separate
project, and modified all my automation for it.