Adding a line "#" in int.sc mode now changes it to "# "
The internal scripting conversion adds and removes lines that only
contain a # to make blank lines work in VVVVVV properly, but that also
means people wouldn't be able to add such a line on purpose without it
getting removed, such as in a text box - as reported in #18.

Actually selectively preserving "#" in text boxes would add more
complexity to the code than I feel is worth it for this edge case,
so I'll leave that aspect WONTFIX, adding a space is a workaround.

Closes #18.