Moved entity displaying to its own function outside displayentities()
That means there's now a possibility to programmatically draw a single
entity, instead of all entities in a room. I'll use this for something
soon. The function is called displayentity.

The function drawentity has been renamed to drawentitysprite to avoid
a bit of confusion. If your plugin uses drawentity, you need to change
that for it to keep working, try something like this:

if drawentitysprite ~= nil then
	drawentity = drawentitysprite
end