Improved FPS limiting (if enabled)
The old FPS limiting method wasn't sophisticated enough, thus limiting
to 60 FPS would instead limit to something closer to 120, limiting to
30 would limit to 57, limiting to 15 would limit it to 29.

The new method accounts for the time actually needed to draw a frame,
and keeps the FPS at an accurate value.

The FPS limit setting has been shifted: instead of limiting to
15, 30, 60 and disabled, the options are now 30, 60, 120 and disabled
respectively, which was closer to what they were actually doing. So if
your FPS limit was set to 30 because that limited it closest to 60,
the option will now display 60 and will limit to 60.