Added handling of OS-dependent forbidden characters in filenames
Ved used to rely on operating system error messages when forbidden
characters are used in filenames. Trying to use a forbidden character
in a filename normally causes failure to write the file which raises
an error message from the OS that Ved displays.

However, for some bizarre reason, Windows lets programs (including Ved)
create a file with a colon (:) in its name without throwing an error,
despite : being a forbidden character and causing obvious bugs,
including the fact that you may need some not-so-straightforward tricks
to even delete the file again, or even its parent directory!

So now Ved itself stops you from using : * ? " < > | in your file paths
if you're on Windows, or : if you're on Mac.