Optimized drawing tiles
SpriteBatches are now used to draw tiles which should be a lot more
efficient and are _the_ way to even do this kind of thing, which I
wasn't really aware of. That means quads for 1200 tiles now are no
longer individually drawn, but added to a spritebatch and then the
spritebatch is drawn. This means higher performance.