Added refresh button to update check/latest info, and scrolling text
So from this version onwards, you can now see changes to the changelog
or see whether a new version has been released without having to
restart Ved.

Scrolling text can now also be displayed under the version info via
the update checking system, which could be used in case it turns out
there's a major issue in a certain version of Ved and I need to warn
users of that version not to save with it, or something. So in
principle, I don't plan on using it much if at all.