Officially bumped minimum LÖVE version from 0.9.0 to 0.9.1
It turns out Ved has been mostly unusable on 0.9.0 for 6.5 months now,
due to a bug in LÖVE itself that was fixed in 0.9.1 ("Fixed
love.graphics.setCanvas() to restore the proper viewport and scissor
rectangles.")

I don't think there's much reason left to keep 0.9.0 supported (except
for me to brag about the wide amount of versions supported). Nobody
depends on it as far as I know (0.9.1 and 0.9.2 have the same system
requirements), and nobody is using it anymore either judging by the
fact I don't remember anybody reporting this for all this time:
https://cdn.discordapp.com/attachments/403718283855134720/555111580250013696/unknown.png

Plus I can now make the code a little cleaner with a few other changes
that were made in 0.9.1.