Settings file cannot be broken with newlines in filenames anymore
When a level that has a CR or LF in its filename and you open that
level (thus becoming a recently opened level in the settings file)
those newline characters are now correctly encoded.

Additionally, errors in the settings file are now caught and you can
choose to reset your settings if it happens. The contents of the
broken settings file are saved in a crash log.

Also, newlines in filenames will now be displayed as question marks
in the levels list because honestly, other options are probably worse.
That is, if newlines actually don't split filenames in parts (#45),
which I'll hopefully get around to soon.

Closes #44.

[pre02]