Migrated part of Windows filesystem code to FFI
FFI basically allows close interaction with, in this case, Windows'
own filesystem functions.

Some of the advantages of doing it this way are that I can get rid of
flashing command prompt windows, and the levels list can show more
info like last modification dates on files.

This is still work in progress, but this is a beginning. **You'll now
see last modification dates on level files if you're on Windows.**

Some drawbacks that'll get fixed:
- Viewing subfolders is temporarily gone
- Time format cannot be chosen yet (is now YYYY-MM-DD 24h)
- Non-ASCII characters will show up even worse than they already did

Testing is appreciated! Other OSes will follow later.

[pre03]