Add adjacent spike lines
If there's a spike right next to the border of the current room, in an
adjacent room, then that will show up on the side as a red line. If the
side is warping then the line will be yellow instead.

Also holy fuck that spike detection code is the ugliest hack I've ever
seen