Add plus-shape locking
Previously, if you were to hold down both [ and ], it would just lock
the cursor in place, which wasn't very useful.

This commit makes it so that when you hold [ and ], you're entering a
mode where you're basically drawing in the plus-shape created by the two
lines. However you have to put your mouse in one of the two lines in
first, in order for Ved to know which line you want to move in, but then
afterwards you can move your mouse out and the cursor will stay locked
in to that line.

The rendering has also been changed so the area that isn't in the
plus-shape isn't darker than if only one of [ or ] had been held, and
the tile where the two lines intersect doesn't stick out.

This closes http://ved.idea.informer.com/proj/?ia=104672