Optimized levels list for large amounts of levels
Not sure if it's the root cause of #48, but the levels list would
display ALL levels in the current directory (even those offscreen)
and more importantly, filtered levels every frame. That's fine if
there are not many levels, but large amounts of levels will start to
cause the performance to worsen.

Now levels are only filtered once every time the input changes, and
offscreen levels are not displayed.

[pre06]