Rewrote Linux filesystem code for FFI
Last modification dates are now shown on Linux, and problems caused
by newlines in filenames can now be resolved. Closes #45.

I don't think I can do this the same way as on Windows (making
assumptions about how structs are structured) because of
non-uniformity of anything categorized as Linux, which you need to
worry less about when writing a C program because, for example,
dirent.d_type being available on one platform and not on the other
won't hopelessly overwrite the wrong memory if you decide not to use
that.
Luckily, you can write a library in C, let each system compile that,
and then use its functions. This does require gcc to be available, but
if it's not, we'll simply fall back to using `ls` as before. If you
want last modification dates, simply install gcc and they'll suddenly
appear next run. (At least, they should!)

For Mac I can probably use the same library, but most Macs probably
don't have any compiler installed. But I can count on uniformity
(and hopefully, uniformity across macOS versions!) So I could compile
it once on my old OS X 10.9 and include that.

[pre13]