Rewrote macOS filesystem code for FFI as well
Last modification dates are now shown on Mac, and actually, one feature
is now available on Mac that was only supported on Windows and Linux
before: subfolders in the levels folder!

This uses the same shared library as I made for Linux, except it's
already compiled. I tested it on 10.9 (Mavericks) and 10.14 (Mojave)
so we should be good, but if it does break for anyone, please tell me.

[pre14]