Added last edit dates to recent levels, title now scrolls if too long
Last edit dates weren't yet added to the list of most recently opened
levels.
Also, if a title in a level's metadata is too long not to overlap with
the last edit date in the levels list, it'll now be cut off and an
arrow will be shown, and when selecting the level (with your keyboard
or by hovering over it) the title will scroll. Titles will only not fit
in the list when they're already two characters longer than what will
fit on screen in VVVVVV itself, so it's not like 60% of your levels
will have their titles crammed in a tight space because of the date.

[pre19]