Int.sc scripts ending with inevitable jump don't get final loadscript
All scripts with internal scripting mode enabled get loadscript(stop)
appended to the end of them, because the script should end when it
ends, there's no reason to want to continue into a certain text box
with scripting commands in it. But it does take up a line, and it may
not be needed if the internal script jumps to another script by itself.
People may mistakenly think they have to include the loadscript(stop)
themselves, which is needed when not using int.sc mode, or they may
simply want to continue in another script.

So if the last line of the script starts with `loadscript(`,
`iftrinkets(0,` or `customiftrinkets(0,`, (accounting for commas or
brackets, but not spaces or additional lines) the loadscript(stop) will
not be included.

[pre20]