Added start of assets viewing functions: a music player!
Actually, I also implemented a music _editor_, because it wasn't really
hard to do, but it might need a bit more polish.

What's still planned is an image viewer and sound effects player.

I'm also thinking it'd be cool to have an easy way to install resource
packs and the like, or easily swap out assets (whose VVVVVV folder
doesn't have tons of suffixed graphics and sounds folders and music
files in it). But I'm not sure about anything there yet.

(I also think we should standardize on some kind of format for adding
textual notes to songs in a custom .vvv, so you could label songs with
song titles and notes for yourself. I might write up a document and ask
others to comment on it, later)

[pre03]