Fix typing pipes (PC) and dollar signs (3DS) in script editor
This makes it so Ved will insert a newline when typing a pipe in non-3DS
mode or when typing a dollar sign in PleaseDo3DSHandlingThanks mode.
This is to hint that pipes and dollar signs simply can't exist as their
own characters inside VVVVVV, and to not be misleading and insert the
pipe or dollar sign anyway, only for it to break in-game.