Fix input_r being ignored in script editor
If you moved the cursor back on a line the syntax highlighter would only
process up to your cursor, making the rest of the line white and making
text in dialog boxes suddenly not be highlighted as dialog boxes
anymore.

This fixes that, and as a nice side bonus you will always be able to see
the character your cursor is on (but during the part of the flash cycle
where the cursor is gone the character will be white instead), unless
you turn syntax highlighting off.

This closes http://ved.idea.informer.com/proj/?ia=105046