Fix Ved's interpretation of non-positive lines
VVVVVV will interpret say/reply,-1 as a 1-line text box, then pass on
the negative for 'text' to deal with, which will then interpret it as a
0-line text box and be able to cause the cutscene-bars internal
scripting exploit. This still happens if the magnitude of the negative
number is bigger than 1.

However, previously Ved didnt't interpret it that way, and would have
put the dialogue box sidelines on the side of the command if it
encountered a negative number. This commit makes it so it'll display it
how VVVVVV will view it.