Add field field 10 for dialog text fields
Field field 10 is basically the input_r if the field is a text field. As
well, all text fields' field 6 is now properly DF.TEXT instead of nil,
and 7-9 are {}, {}, and "", because Lua and unpack() in particular do
not like nils. I saw field 9 being used in dialog.form.exportmap_make()
and it kind of looked like a string, so I made it a string.

This is to be able to move the cursor in dialog text fields.