Add permanent keyboard subtool select via holding SHIFT
The subtool keybinds go like Z, X, C, V, H, B, and F. You can hold them
down to temporarily switch to them, but if you let go you would revert
back to the tool you previously held.

Previously, there was no other way to make the keyboard subtool switch
"stick", i.e. stay permanent when you let go of the relevant key. You
had to use the scrollwheel instead (or you could click on it, or use
CTRL/CMD+</>).

This commit makes it so that if you're holding SHIFT when you press the
key, the subtool will stick when you let go of the key. This is useful
except for the custom brush size tool in tools 1 and 2 (wall and
background), which currently does not have a keybind and must be
scrolled to (or clicked on, or scrolled to with CTRL/CMD+</>).