Allow canceling script entity (re)naming by pressing ESC
Basically, you place a terminal or a trigger, but then get to the naming
part and realize you don't want to rename it, so you'd press ESC.  Or
you rename an existing terminal or trigger, only for you to realize you
don't actually want to rename it. In both cases, pressing ESC will now
let you cancel the (re)name action.

There are two bugs with this:
 - Canceling a newly-placed script entity will internally delete it,
   unnecessarily dirtying - but then again, naming a newly-placed script
   entity to be blank will do that, too
 - Canceling a rename of a script entity already named "stop" will
   delete it instead of leaving it be. File this under the growing list
   of "reasons why Ved should discontinue using 'stop' as a temporary
   internal placeholder"