Changed script box/terminal/roomtext placeholders to empty strings
Script boxes and terminals used to be initialized with "stop" or L.BUG
as their data, now they're just empty strings. This makes the code for
!15 better, not only because it doesn't need to check for entity types
anymore to determine what the placeholder is, but also because it's
more likely that you'd want to name your script "stop" or have roomtext
with "[Bug!]" than that you'd want an empty string as script name or
roomtext.

[pre09]