Added implementation of extra metadata for .vvv music
This uses the format we discussed on Discord, which is documented here:
https://tolp.nl/vvvmusic/

That means you can now read and write:
- Song titles
- Song filenames (automatically set when loading a song in, but can be
  changed)
- Song notes
- File export date
- File/Album title
- Artist
- File notes

The title _or_ filename for each song is displayed in the music
player/editor when available.
Just to be clear, this has nothing to do with OGG metadata or anything
like that. You can save a .vvv with or without the extra metadata.

I just need to create a multiline input field for the notes fields.
Also, the file export date doesn't yet get updated in the opened file
when you export it, but it is written to the file.

[pre10]