Holding name changed from 1.5.2 to 1.6.0.
After this the latest intermediate versions will be:
1.6.0-pre11 (without Assets and music editor, this will be the "latest
            intermediate version")
1.6.0-pre12 (with Assets and music editor, this will say you're not
            up-to-date but you'd know)

Until everything changes again. But basically, I'm keeping the
intermediate version number for the assets_viewer branch ahead of the
master branch, and have the master branch version be indicated within
Ved as "latest". This system is kind of chaos when you work in
separate branches and don't always immediately merge master into other
branches, but all it takes to fix that is to change the number to a
higher one than it has ever been before and make everything older than
that outdated anyway.

[pre11]