Fix Ctrl+D on the last line of a script
If you pressed Ctrl+D on the last line of a script, it'd appear to blank
the entire line instead of removing it entirely, and then you'd have to
press backspace twice instead of once to remove the line and go to the
previous line (but moving up and then back down, or clicking on the
not-yet-missing line which would move the cursor to the line above so
you could click again to the last line, and then pressing backspace to
remove the line, would properly remove the line on the first press).

This is the behavior I mentioned in the third edit of #55, and this
fixes that buggy behavior.