Allow Delete to concat/del lines&tab to level input
The last part is talking about being able to press Backspace while
having selected another level in level select to deselect the level, and
this commit makes it so that can happen as well with Delete.

Pressing Delete can now combine/remove lines in the script editor and in
the level notepad. To elaborate, if you have lines like this, and your
cursor is at the underscore:

    line 1
    _
    line 2

Pressing Delete will now delete the line you're on and put your cursor
to the next line like so:

    line 1
    _ine 2

Previously, it would have done nothing.

You can also press Delete to combine lines while reducing line count. If
you have your cursor at the end of a line, like so:

    line 1_
    line 2

Pressing Delete will combine the two lines and leave your cursor in
place, like so:

    line 1_ine 2

Previously, it would have done nothing.