Add Ctrl+Shift+D to remove lines, moving upwards
This is to mirror the use-case of holding down Ctrl+D to delete a bunch
of lines, moving downwards. Now you can use Ctrl+Shift+D and hold it
down to delete a bunch of lines, moving upwards.

I'm doing this in the same branch as fixing Ctrl+D on the last line of a
script so we won't have to deal with a merge conflict if I did it in
another branch.