Merge textlinestogo and textlinestogo2
This fixes a bug where something like this could cause the dialogue box
(except the line on the side) to be highlighted in an alt color:

    say,5
    text,,,,3
    [line]
    [line]
    [line]
    createcrewman

This is because textlinestogo2 didn't check that it was already inside a
text box of say/reply, which textlinestogo would have checked already.
Another solution (to keep the two variables separate) would be to
account for say/reply in textlinestogo2's system, but that would be very
wasteful as it would duplicate work textlinestogo had already done.