Fix a critical bug with flag swapping 1-digits affecting 2-digits
This was because the flag command detection code didn't bother checking
for another argument separator and just assumed that
<command><separator><flag> was enough. However, the flag could be a
single digit, which would change all the other two-digit flags if the
first digit of those two-digit flags was the single digit.

For a concrete example:

We're going to (horribly) swap flags 1 and 2.

The flag swapping code would detect this command:

    flag(20,on)

And the part it would get is "flag(2", since it didn't bother checking
for another argument separator. Then, it would replace "flag(2" with
"flag(1", mucking up the usage of flag 20 that is nowhere near relevant
to flags 1 or 2:

    flag(10,on)

Thus your flags would get mangled up if you swapped any single-digit
flags at all.

Please merge this IMMEDIATELY, @Dav999.